Move the `QAction.setShortcut` patch to `QtGui`
