+++ qtpy/QtGui.py
@@ -8,7 +8,11 @@
 
 """Provides QtGui classes and functions."""
 
-from . import PYQT5, PYQT6, PYSIDE2, PYSIDE6, QtModuleNotInstalledError
+
+from packaging.version import parse
+
+from . import PYQT5, PYQT6, PYSIDE2, PYSIDE6
+from . import QT_VERSION as _qt_version
 from ._utils import getattr_missing_optional_dep, possibly_static_exec
 
 _missing_optional_names = {}
@@ -252,3 +256,47 @@ if PYQT6 or PYSIDE6:
     # Follow similar approach for `QDropEvent` and child classes
     QDropEvent.pos = lambda self: self.position().toPoint()
     QDropEvent.posF = lambda self: self.position()
+
+
+# Make `QAction.setShortcut` and `QAction.setShortcuts` compatible with Qt>=6.3
+if PYQT5 or PYSIDE2 or parse(_qt_version) < parse("6.3"):
+    from functools import partialmethod
+
+    from ._utils import (
+        set_shortcut,
+        set_shortcuts,
+    )
+
+    class _QAction(QAction):
+        old_set_shortcut = QAction.setShortcut
+        old_set_shortcuts = QAction.setShortcuts
+
+        def setShortcut(self, shortcut):
+            return set_shortcut(
+                self,
+                shortcut,
+                old_set_shortcut=_QAction.old_set_shortcut,
+            )
+
+        def setShortcuts(self, shortcuts):
+            return set_shortcuts(
+                self,
+                shortcuts,
+                old_set_shortcuts=_QAction.old_set_shortcuts,
+            )
+
+    _action_set_shortcut = partialmethod(
+        set_shortcut,
+        old_set_shortcut=QAction.setShortcut,
+    )
+    _action_set_shortcuts = partialmethod(
+        set_shortcuts,
+        old_set_shortcuts=QAction.setShortcuts,
+    )
+    QAction.setShortcut = _action_set_shortcut
+    QAction.setShortcuts = _action_set_shortcuts
+    if (  # despite the two previous lines!
+        QAction.setShortcut is not _action_set_shortcut
+        or QAction.setShortcuts is not _action_set_shortcuts
+    ):
+        QAction = _QAction
